requirements.txt
langchain
langchain_community
langgraph
langchain-core
langchain-openai
langgraph-checkpoint-sqlite

python-dotenv
langsmith
langgraph-cli[inmem]
bs4

fastapi
passlib
sqlmodel
itsdangerous
jinja2
python-multipart
bcrypt